remove user from My Favourites
